refactor(tests): split behavioral tests to comply with LOC limits
Split test_check_sync_rules.py into two files:
- test_check_sync_rules.py (192 lines): config-only tests + 2 existing
  behavioral tests
- test_check_sync_rules_behavioral.py (279 lines): 4 new behavioral tests
  for sync rule guards

Both files now under 400 line CI block threshold.
All 9 tests still pass.

Refactoring triggered by LOC check during vibe-commit workflow.
